#256448: ecl1 now handles linked work trees gracefully.
Also: Auto lfs prune now runs in its own thread. I suspect that this was
the reason why sometimes other earlyStartup extension were not started.
I also strongly suspect that this fixes #255994

+++ net.sf.ecl1.git.updatehooks/src/net/sf/ecl1/git/updatehooks/UpdateHooks.java
@@ -9,6 +9,7 @@ import java.nio.file.StandardCopyOption;
 import java.util.Map.Entry;
 import java.util.TreeMap;
 
+import org.eclipse.core.resources.IFolder;
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.IWorkspaceRoot;
 import org.eclipse.core.resources.ResourcesPlugin;
@@ -63,15 +64,18 @@ public class UpdateHooks implements IStartup {
 				 */
 				IProject webapps = WebappsUtil.findWebappsProject();
 				if(webapps != null) {
-					try {
-						Path hooksDirWebapps = Paths.get(webapps.getFolder(HOOKS_DIR_ECLIPSE_PROJECTS).getLocationURI());
-
-						Files.copy(getCommitMsgHook(), hooksDirWebapps, StandardCopyOption.REPLACE_EXISTING);
-						logger.info("Successfully updated the git hooks of the webapps project.");
-					} catch (IOException e) {
-						logger.error2("Failed to update the git hooks of the webapps project. Exception: " + e.getMessage(), e);
+					if(webapps.getLocation().append(".git").toFile().isFile()) {
+						logger.info("Current project is managed by git, but you are currently in a linked work tree. Updating the git hooks will not work in a linked work tree. Skipping...");
+					} else {
+						try {
+							Path hooksDirWebapps = Paths.get(webapps.getFolder(HOOKS_DIR_ECLIPSE_PROJECTS).getLocationURI());
+	
+							Files.copy(getCommitMsgHook(), hooksDirWebapps, StandardCopyOption.REPLACE_EXISTING);
+							logger.info("Successfully updated the git hooks of the webapps project.");
+						} catch (IOException e) {
+							logger.error2("Failed to update the git hooks of the webapps project. Exception: " + e.getMessage(), e);
+						}
 					}
-					
 				} else {
 					logger.info("No webapps project found in your workspace.");
 				}
